importmodule

+++ 0x02-python-import_modules/calculator_1.py
@@ -7,7 +7,7 @@ def sub(a, b):
 
 
 def mul(a, b):
-    return a * b
+    return float(a * b)
 
 
 def div(a, b):
